moved commandActionBuilder from helpers.ts to base.ts
