+++ test/e2e/commands/separate_node_update.test.ts
@@ -28,7 +28,6 @@ import {
   getTmpDir,
   HEDERA_PLATFORM_VERSION_TAG,
 } from '../../test_util.js';
-import {getNodeLogs} from '../../../src/core/helpers.js';
 import {HEDERA_HAPI_PATH, MINUTES, ROOT_CONTAINER} from '../../../src/core/constants.js';
 import fs from 'fs';
 import type {PodName} from '../../../src/types/aliases.js';
@@ -66,7 +65,7 @@ e2eTestSuite(namespace, argv, undefined, undefined, undefined, undefined, undefi
     after(async function () {
       this.timeout(10 * MINUTES);
 
-      await getNodeLogs(k8, namespace);
+      await k8.getNodeLogs(namespace);
       await nodeCmd.handlers.stop(argv);
       await k8.deleteNamespace(namespace);
     });
